Switch type for amount from BIGINT to MONEY

+++ MoneyCheckWebApp/Types/Purchases/PurchaseType.cs
@@ -9,7 +9,7 @@ namespace MoneyCheckWebApp.Types.Purchases
     {
         public long Id { get; set; }
         public DateTime BoughtAt { get; set; }
-        public long Amount { get; set; }
+        public decimal Amount { get; set; }
         public long CategoryId { get; set; }
         public decimal? Longitude { get; set; }
         public decimal? Latitude { get; set; }
